[REFACTOR] 면접만들기 관련 atom interviewMakeAtom으로 리팩토링
